requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==4.2.7
djangorestframework==3.14.0
idna==3.11
pillow==12.0.0
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
gunicorn==23.0.0
whitenoise==6.6.0
dj-database-url==1.3.0
